+++ pyvo/registry/regtap.py
@@ -487,8 +487,7 @@ class RegistryResource(dalq.Record):
         (f"\n  ivo_string_agg(COALESCE(intf_type, ''), '{TOKEN_SEP}')",
             "intf_types"),
         (f"\n  ivo_string_agg(COALESCE(intf_role, ''), '{TOKEN_SEP}')",
-            "intf_roles"),
-        "alt_identifier"]
+            "intf_roles")]
 
     def __init__(self, results, index, session=None):
         dalq.Record.__init__(self, results, index, session=session)
@@ -583,14 +582,6 @@ class RegistryResource(dalq.Record):
         URL pointing to a human-readable document describing this resource.
         """
         return self.get("reference_url", decode=True)
-
-    @property
-    def alt_identifier(self):
-        """Alternative identifier.
-
-        It is often used to provide the resource associated DOI.
-        """
-        return self.get("alt_identifier", decode=True)
 
     @property
     def creators(self):
@@ -935,8 +926,14 @@ class RegistryResource(dalq.Record):
                 else:
                     print(f"Authors: {', '.join(creators[:nmax_authors])} et al.\n"
                     "See creators attribute for the complete list of authors.", file=file)
-            if self.alt_identifier:
-                print(f"Alternative identifier: {self.alt_identifier}", file=file)
+
+            alt_identifiers = self.get_alt_identifiers()
+            if alt_identifiers:
+                print(
+                    "Alternative identifier(s): {}".format(
+                        ", ".join(alt_identifiers)),
+                    file=file)
+
             if self.reference_url:
                 print("More info: " + self.reference_url, file=file)
 
@@ -964,6 +961,17 @@ class RegistryResource(dalq.Record):
             contacts.append(contact)
 
         return "\n".join(contacts)
+
+    def get_alt_identifiers(self):
+        """return a sequence of non-ivoid identifiers for the resource.
+
+        This is typically used to provide a DOI for the resource.
+        """
+        res = get_RegTAP_service().run_sync("""
+            SELECT alt_identifier
+            FROM rr.alt_identifier
+            WHERE ivoid={}""".format(rtcons.make_sql_literal(self.ivoid)))
+        return [r["alt_identifier"] for r in res]
 
     def _build_vosi_column(self, column_row):
         """